adding all approval

+++ blogger/src/main/java/com/rayhan/blogger/repository/BloggerPostRepository.java
@@ -6,7 +6,9 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.rayhan.blogger.entity.BloggerPost;
 
-public interface BloggerPostRepository extends JpaRepository<BloggerPost, Integer> {
+public interface BloggerPostRepository extends JpaRepository<BloggerPost, Long> {
+	
+	
 	@Query(value="SELECT  id FROM login_activity WHERE login_activity.token=:s",nativeQuery = true)    
 	int findUserId(String s);
